added the get_string function to extract text from images in another file to make the code look better
